clean up API
Removed old methods from API, and split off toMap primitive

Needs to re-introduce the old methods (single array and whatever that
calendar thing is)
