+++ src/main/scala/data/ApiResultF.scala
@@ -8,7 +8,7 @@ import cats.implicits._
 case class APIResultF[A, B](count: Int, next: Option[B], previous: Option[Uri], results: List[A])
 
 object APIResultF {
-  implicit def apiResultFunctor: Bitraverse[APIResultF] = new Bitraverse[APIResultF] {
+  implicit val apiResultFunctor: Bitraverse[APIResultF] = new Bitraverse[APIResultF] {
     override def bimap[A, B, C, D](fab: APIResultF[A, B])(f: (A) => C, g: (B) => D): APIResultF[C, D] = {
       val newNext    = fab.next.map(g)
       val newResults = fab.results.map(f)
